C99 compatibility fix
I researched this quite a bit, and this approach seems to be the least
bad option.

Future compilers are likely to require functions to be declared before
they can be called.  The flex lexer generator can be taught to create a
header file with the required declarations.  However, automake conspires
against us to use it during the build.  So I extracted the relevant
declarations by hand and put it into a non-generated header file.  At
least the declarations are checked against the definitions in
rec-sex-lex.c, so maybe it's not too bad after all.

Related to:

  <https://fedoraproject.org/wiki/Changes/PortingToModernC>
  <https://fedoraproject.org/wiki/Toolchain/PortingToModernC>